refactor: split campaign production contract
